backend_python/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
